+++ src/helpers/getExcludeYourselfFromTeamMessage.ts
@@ -4,7 +4,7 @@ export const getExcludeYourselfFromTeamMessage = (
     currentUserLogin?: string
 ) => {
     if (userName === currentUserLogin) {
-        return `You are trying to exclude yourself from the "${teamName}". If you do this, you will no longer be able to see this team. Are you sure you want to exclude yourself?`;
+        return `You are trying to exclude yourself from the "${teamName}". If you do this, you will no longer be able to access this team. Are you sure you want to exclude yourself?`;
     }
     return `Exclude "${userName}" from "${teamName}"?`;
 };
